Cursor Update
Fixed animation glitch when I click on the course cards for the drilldown where the screen/ui would glitch a little during the opening animation. This was only happening when the UI initially loaded the csv and was only present when there were a lot of course cards present (example: when the unfiltered Cal_GETC_Areas pathway is selected).